Добавит дополнительные модули
