+++ index.html
@@ -35,8 +35,8 @@
           </select>
           <select name="housing-price" id="housing-price" class="map__filter">
             <option value="any" selected>Любая</option>
-            <option value="middle">10000 - 50000&#x20bd;</option>
             <option value="low">до 10000&#x20bd;</option>
+            <option value="middle">10000 - 50000&#x20bd;</option>
             <option value="high">от 50000&#x20bd;</option>
           </select>
           <select name="housing-rooms" id="housing-rooms" class="map__filter">
@@ -47,8 +47,8 @@
           </select>
           <select name="housing-guests" id="housing-guests" class="map__filter">
             <option value="any" selected>Любое число гостей</option>
-            <option value="2">Два гостя</option>
             <option value="1">Один гость</option>
+            <option value="2">Два гостя</option>
             <option value="0">Не для гостей</option>
           </select>
           <fieldset id="housing-features" class="map__features">
@@ -133,9 +133,9 @@
         <fieldset class="ad-form__element">
           <label class="ad-form__label" for="capacity">Количество мест</label>
           <select id="capacity" name="capacity">
-            <option value="3">для 3 гостей</option>
-            <option value="2">для 2 гостей</option>
             <option value="1" selected>для 1 гостя</option>
+            <option value="2">для 2 гостей</option>
+            <option value="3">для 3 гостей</option>
             <option value="0">не для гостей</option>
           </select>
         </fieldset>
