+++ js/form.js
@@ -82,68 +82,4 @@ timeForm.addEventListener('change', (evt) => {
   timeOut.value = evt.target.value;
 });
 
-const showUploadSuccessAlert = () => {
-  const successAlert = document.querySelector('#success').content.querySelector('.success');
-  document.body.append(successAlert);
-
-  document.addEventListener('keydown', (evt) => {
-    if (evt.key === 'Escape' || evt.key === 'Esc') {
-      evt.preventDefault();
-      successAlert.remove();
-    }
-  });
-
-  document.addEventListener('click', (evt) => {
-    evt.preventDefault();
-    successAlert.remove();
-  });
-};
-
-const showUploadErrorAlert = () => {
-  const errorAlert = document.querySelector('#error').content.querySelector('.error');
-  document.body.append(errorAlert);
-
-  document.addEventListener('keydown', (evt) => {
-    if (evt.key === 'Escape' || evt.key === 'Esc') {
-      evt.preventDefault();
-      errorAlert.remove();
-    }
-  });
-
-  document.addEventListener('click', (evt) => {
-    evt.preventDefault();
-    errorAlert.remove();
-  });
-
-  const errorButton = document.querySelector('.error__button');
-  errorButton.addEventListener('click', (evt) => {
-    evt.preventDefault();
-    errorAlert.remove();
-  });
-};
-
-const setAdFormSubmit = (onSuccess) => {
-  adForm.addEventListener('submit', (evt) => {
-    evt.preventDefault();
-
-    const formData = new FormData(evt.target);
-
-    fetch('https://24.javascript.pages.academy/keksobooking1', {
-      method: 'POST',
-      body: formData,
-    })
-      .then((response) => {
-        if (response.ok) {
-          onSuccess();
-          showUploadSuccessAlert();
-        } else {
-          showUploadErrorAlert();
-        }
-      })
-      .catch(() => {
-        showUploadErrorAlert();
-      });
-  });
-};
-
-export { togglePageState, adForm, mapFilters, setAdFormSubmit };
+export { togglePageState, adForm, mapFilters };
